requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
python-dotenv==1.0.1
groq==0.4.2
azure-cognitiveservices-speech==1.34.1
aiofiles==23.2.1
pydub==0.25.1
